started progressing UI
Made some changes to Patient class
added a oldSickness attribute.

created Colors class, its not important but it allows us to use multiple colors in the ui, explanation on how to use them is written inside the class.

Started working on a functional interface in Main class, try it and see how it goes but it still needs alot of work.